Added filtering and fixed bugs
Removed the slider minimum price JFormatedTextField
Moved the maximum price JFormatedTextField above of the slider
Now can be searched words in the the JTextArea components of description
and the address located in the cell 
Added 2 filter options:
 - By price: As parameters needs a minimum and maximum price.
 - By matching text: As parameter needs the text that want be searched
in the description and the address.
Renamed the inner class TableDetailsCell to TableCellComponents
Now the inner class CellDetails, stores information of which
TableCellComponents components contains.
Added new key binding to use the test classes to draw component borders