Changes in RuralHouse
Related to the changes on the commit d0e63f4, the variable of the
review, instead of using a int type to store it, now uses an object of
type "Review".
For this reason, now have been added a relation of One to One. The
parameters in the tag @OneToOne, adds the next behaviour: When a review
or a RuralHouse is removed, the one related to is also removed, a non
optional parameters is added to force to have always one review added.

Re-generated the serial version ID

Added constant that store a image to use when no other is available